Schematic API Revamp (#366)
* work on migrating schematic code

* almost finished converting all code to new api

* tests compiling, issues with unconnected ports

* fix tests

* fix dout buffer issue

* fix potential issue

* fix issues with schematic, still issues with layout

* fix tests and lvs

+++ sramgen/src/config/wmask_control.rs
@@ -3,6 +3,6 @@ use crate::config::gate::AndParams;
 #[derive(Debug, Clone)]
 pub struct WriteMaskControlParams {
     pub name: String,
-    pub width: i64,
+    pub width: usize,
     pub and_params: AndParams,
 }
